+++ bin/awsmobile-configure
@@ -23,7 +23,7 @@ const options = ['aws', 'project']
 
 program
     .description('edit configuration settings of the aws account or the project')
-    .usage('[aws|project] [options]')
+    .usage('<aws|project> [options]')
     .option('-l, --list', 'list the current configuration settings')
     .option('-p, --profile <name>', '(aws) use named profile for aws access')
     .option('-n, --new-user', '(aws) walk through the steps to setup aws access')
