+++ lib/aws-operations/aws-client.js
@@ -15,7 +15,6 @@ const AWS = require('aws-sdk')
 
 const awsmobileJSConstant = require('../utils/awsmobilejs-constant.js')
 const cliConfigManager = require('../utils/cli-config-manager.js')
-const awsConfigFileManager = require('./aws-config-file-manager.js')
 
 function Mobile(awsDetails){
     AWS.config.loadFromPath(awsDetails.info.AWSConfigFilePath)
