+++ lib/aws-operations/aws-config-info-manager.js
@@ -17,6 +17,7 @@ const moment = require('moment')
 const lineByLine = require('n-readlines')
 
 const awsMobileRegions = require('./aws-regions.js').regions
+const awsConfigHelper = require('./aws-config-helper.js')
 const projectInfoManager = require('../project-info-manager.js')
 const awsmobileJSConstant = require('../utils/awsmobilejs-constant.js')
 const nameManager = require('../utils/awsmobilejs-name-manager.js')
@@ -36,43 +37,45 @@ function resolve(silentFlag){
 }
 
 function setProfile(profileName){
-    ensureFolderStructure()
-    let awsInfoFilePath = pathManager.getGeneralAWSInfoFilePath()
-
-    let projectInfo = projectInfoManager.getProjectInfo(true)
-    if(projectInfo){
-        awsInfoFilePath = pathManager.getAWSInfoFilePath(projectInfo.ProjectPath)
-    }
-
-    let awsInfo = awsInfoTemplate
-    let executeSet = true
-
-    if(fs.existsSync(awsInfoFilePath)){
-        awsInfo = JSON.parse(fs.readFileSync(awsInfoFilePath, 'utf8'))
-        if(awsInfo.IsUsingProfile && awsInfo.ProfiletName == profileName){
-            executeSet = false
-        }
-    }
-
-    if(executeSet){
-        awsInfo.IsUsingProfile = true
-        awsInfo.ProfiletName = profileName
-        awsInfo.AWSInfoFilePath = awsInfoFilePath
-        if(!awsInfo.AWSConfigFilePath){
-            awsInfo.AWSConfigFilePath = path.normalize(path.join(pathManager.getSysProjectAWSConfigDirPath(), nameManager.generateAWSConfigFileName(projectInfo)))
-        }
-        awsInfo.LastProfileSyncTime = ""
-        let jsonString = JSON.stringify(awsInfo, null, '\t')
-        fs.writeFileSync(awsInfoFilePath, jsonString, 'utf8')
-       
-        if(projectInfo){
-            awsInfo = resolve_project_awsConfig(projectInfo)
-        }else{
-            awsInfo = resolve_general_awsConfig()
-        }
-    }
+    let isSuccessful = true
+    let awsConfig = awsConfigHelper.getSystemConfig(awsInfo.ProfiletName, silentFlag)
+    if(awsConfig){
+        ensureFolderStructure()
+        let awsInfoFilePath = pathManager.getGeneralAWSInfoFilePath()
     
-    return awsInfo
+        let projectInfo = projectInfoManager.getProjectInfo(true)
+        if(projectInfo){
+            awsInfoFilePath = pathManager.getAWSInfoFilePath(projectInfo.ProjectPath)
+        }
+
+        let awsInfo = awsInfoTemplate
+        let executeSet = true
+        if(fs.existsSync(awsInfoFilePath)){
+            awsInfo = JSON.parse(fs.readFileSync(awsInfoFilePath, 'utf8'))
+            if(awsInfo.IsUsingProfile && awsInfo.ProfiletName == profileName){
+                executeSet = false
+            }
+        }
+    
+        if(executeSet){
+            awsInfo.IsUsingProfile = true
+            awsInfo.ProfiletName = profileName
+            awsInfo.AWSInfoFilePath = awsInfoFilePath
+            if(!awsInfo.AWSConfigFilePath){
+                awsInfo.AWSConfigFilePath = path.normalize(path.join(pathManager.getSysProjectAWSConfigDirPath(), nameManager.generateAWSConfigFileName(projectInfo)))
+            }
+            awsInfo.LastProfileSyncTime = moment().format(awsmobileJSConstant.DateTimeFormatString) 
+            let jsonString = JSON.stringify(awsInfo, null, '\t')
+            fs.writeFileSync(awsInfoFilePath, jsonString, 'utf8')
+
+            let jsonString = JSON.stringify(awsConfig)
+            fs.writeFileSync(awsConfigFilePath, jsonString, 'utf8')
+        }
+    }else{
+        isSuccessful = false
+    }
+
+    return isSuccessful
 }
 
 function setNoProfileSync(awsInfo){ 
@@ -88,12 +91,13 @@ function setNoProfileSync(awsInfo){
 function resolve_general_awsConfig(silentFlag){
     let awsInfoFilePath = pathManager.getGeneralAWSInfoFilePath()
     let awsConfigFilePath = pathManager.getGeneralAWSConfigFilePath()
+
     if(!fs.existsSync(awsInfoFilePath)){
         let awsInfo = awsInfoTemplate
-        if(fs.existsSync(awsConfigFilePath)){
+        if(fs.existsSync(awsConfigFilePath)){ //meaning it was generated by older version of the cli
             awsInfo.IsUsingProfile = false
             awsInfo.ProfiletName = ""
-        }else{ //for backward compatibility
+        }else{ //meaning this usage is started from fresh
             awsInfo.IsUsingProfile = true
             awsInfo.ProfiletName = "default"
         }
@@ -169,9 +173,11 @@ function resolve_project_awsConfig(projectInfo, silentFlag){
 }
 
 function syncProfile(awsInfo, silentFlag){
-    awsConfig = getSystemAwsConfig(awsInfo.ProfiletName, silentFlag)
-    let jsonString = JSON.stringify(awsConfig)
-    fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
+    let awsConfig = awsConfigHelper.getSystemConfig(awsInfo.ProfiletName, silentFlag)
+    if(awsConfig){
+        let jsonString = JSON.stringify(awsConfig)
+        fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
+    }
 }
 
 function isSyncProfileNeeded(awsInfo){
@@ -208,8 +214,8 @@ function isSyncProfileNeeded(awsInfo){
 /////////////////////////////////
 
 const awsInfoTemplate = {
-    "IsUsingProfile": true,
-    "ProfiletName": "default",
+    "IsUsingProfile": false,
+    "ProfiletName": "",
     "AWSConfigFilePath": "",
 	"AWSInfoFilePath": "",
 	"LastProfileSyncTime": "2018-01-01-01-01-01"
@@ -228,160 +234,7 @@ function ensureFolderStructure(){
     fs.ensureDirSync(sysProjectAWSConfigDirPath)
 }
 
-function getSystemAwsConfig(profileName, silentFlag){
-    awsConfig = new Object()
-    let sysCredential = readSystemAwsCredential(profileName, silentFlag)
-    if(sysCredential){
-        awsConfig.accessKeyId = sysCredential.aws_access_key_id ? sysCredential.aws_access_key_id : awsmobileJSConstant.DefaultAWSAccessKeyId
-        awsConfig.secretAccessKey = sysCredential.aws_secret_access_key ? sysCredential.aws_secret_access_key : awsmobileJSConstant.DefaultAWSSecretAccessKey
-    }else{
-        awsConfig.accessKeyId = awsmobileJSConstant.DefaultAWSAccessKeyId
-        awsConfig.secretAccessKey = awsmobileJSConstant.DefaultAWSSecretAccessKey
-    }
 
-    let sysConfig = readSystemAwsConfig(profileName, silentFlag)
-
-    if(sysConfig){
-        awsConfig.region = sysConfig.region ? sysConfig.region : awsmobileJSConstant.DefaultAWSRegion
-    }else{
-        awsConfig.region = awsmobileJSConstant.DefaultAWSRegion
-    }
-    return awsConfig
-}
-
-function readSystemAwsConfig(profileName, silentFlag)
-{
-    let result
-
-    if(!profileName){
-        profileName = 'default'
-    }
-    let profileMark = '[' + profileName + ']'
-
-    if(profileName != 'default'){
-        profileMark = '[profile ' + profileName + ']'
-    }
-
-    let sysConfigFilePath = pathManager.getSysAwsConfigFilePath()
-
-    if(fs.existsSync(sysConfigFilePath)){
-        try{
-            let liner = new lineByLine(sysConfigFilePath)
-            let line
-            let isInProfile = false
-            while (line = liner.next()) {
-                let strPair = line.toString().split("=")
-                if(strPair.length == 1){
-                    if(line == profileMark){
-                        if(!result){
-                            result = {}
-                        }
-                        isInProfile = true
-                    }else{
-                        if(result){
-                            break
-                        }
-                        isInProfile = false
-                    }
-                }else if (strPair.length == 2){
-                    if(isInProfile){
-                        result[strPair[0].trim()] = strPair[1].trim()
-                    }
-                }
-            }
-            if(!result){
-                if(!silentFlag){
-                    console.log()
-                    console.log(chalk.red('no config found for profile ' + profileName))
-                    console.log('aws-cli config file queried:')
-                    console.log(sysConfigFilePath)
-                }
-            }
-        }catch(e){
-            if(!silentFlag){
-                console.log()
-                console.log(chalk.red('error on reading config for profile ' + profileName))
-                console.log('aws-cli config file queried:')
-                console.log(sysConfigFilePath)
-                result = undefined
-            }
-        }
-    }else{
-        if(!silentFlag){
-            console.log()
-            console.log(chalk.red('no aws-cli config file is found'))
-            console.log('aws-cli config file queried:')
-            console.log(sysConfigFilePath)
-        }
-    }
-
-    return result
-}
-
-function readSystemAwsCredential(profileName, silentFlag)
-{
-    let result
-    
-    if(!profileName){
-        profileName = 'default'
-    }
-    let profileMark = '[' + profileName + ']'
-
-    let sysCredentialsFilePath = pathManager.getSysAwsCredentialsFilePath()
-
-    if(fs.existsSync(sysCredentialsFilePath)){
-        try{
-            let liner = new lineByLine(sysCredentialsFilePath)
-            let line
-            let isInProfile = false
-            while (line = liner.next()) {
-                let strPair = line.toString().split("=")
-                if(strPair.length == 1){
-                    if(line == profileMark){
-                        if(!result){
-                            result = {}
-                        }
-                        isInProfile = true
-                    }else{
-                        if(result){
-                            break
-                        }
-                        isInProfile = false
-                    }
-                }else if (strPair.length == 2){
-                    if(isInProfile){
-                        result[strPair[0].trim()] = strPair[1].trim()
-                    }
-                }
-            }
-            if(!result){
-                if(!silentFlag){
-                    console.log()
-                    console.log(chalk.red('no credentials found for profile ' + profileName))
-                    console.log('aws-cli credentials file querried:')
-                    console.log(sysCredentialsFilePath)
-                }
-            }
-        }catch(e){
-            if(!silentFlag){
-                console.log()
-                console.log(chalk.red('error on reading credentials for profile ' + profileName))
-                console.log('aws-cli credentials file querried:')
-                console.log(sysCredentialsFilePath)
-                console.log(e)
-                result = undefined
-            }
-        }
-    }else{
-        if(!silentFlag){
-            console.log()
-            console.log(chalk.red('no aws-cli credentials file is found'))
-            console.log('aws-cli credentials file querried:')
-            console.log(sysCredentialsFilePath)
-        }
-    }
-    return result
-}
 
 /////////////////////////////////////////////
 function validateAWSConfig(awsConfig){
