+++ lib/aws-operations/aws-config-manager.js
@@ -18,7 +18,7 @@ const lineByLine = require('n-readlines')
 const util = require('util')
 
 const awsMobileRegions = require('./aws-regions.js').regions
-const awsConfigFileManager = require('./aws-config-file-manager.js')
+const awsConfigFileManager = require('./aws-config-info-manager.js')
 const awsmobileJSConstant = require('../utils/awsmobilejs-constant.js')
 const pathManager = require('../utils/awsmobilejs-path-manager.js')
 const nameManager = require('../utils/awsmobilejs-name-manager.js')
@@ -44,7 +44,7 @@ function checkAWSConfig(callback){
                 }
             ]).then(function (answers) {
                 if(answers.configureAWS){
-                    configureAWS(null, function(awsDetails){
+                    setConfig(function(awsDetails){
                         if(awsConfigFileManager.validateAWSConfig(awsDetails.config)){
                             callback(awsDetails)
                         }else{
@@ -192,72 +192,81 @@ function PressAnyKeyToContinue(message, callback){
     })
 }
 
-function configureAWS(profileName, callback){
-    if(profileName){
-        console.log('setting awsmobile-cli to use named profile: ' + chalk.blue(profileName))
-        let awsInfo = awsConfigFileManager.setProfile(profileName) 
-        let awsConfig = JSON.parse(fs.readFileSync(awsInfo.AWSConfigFilePath, 'utf8'))
-        if(callback){
-            callback(constructAwsDetails(awsInfo, awsConfig))
+function setConfig(callback){
+    let awsDetails = resolveAWSConfig()
+    let awsInfo = awsDetails.info
+    let awsConfig = awsDetails.config
+
+    inquirer.prompt([
+        {
+            type: 'input',
+            name: 'accessKeyId',
+            message: "accessKeyId: ",
+            default: awsConfig.accessKeyId
+        },
+        {
+            type: 'input',
+            name: 'secretAccessKey',
+            message: "secretAccessKey: ",
+            default: awsConfig.secretAccessKey
+        },
+        {
+            type: 'list',
+            name: 'region',
+            message: "region: ",
+            choices: awsMobileRegions,
+            default: awsConfig.region
         }
+    ]).then(function (answers) {
+        let awsConfigChanged = false
+        if(answers.accessKeyId){
+            let newKeyId = answers.accessKeyId.trim()
+            if(awsConfig.accessKeyId != newKeyId){
+                awsConfig.accessKeyId = newKeyId
+                awsConfigChanged = true
+            }
+        }
+        if(answers.secretAccessKey){
+            let newKey = answers.secretAccessKey.trim()
+            if( awsConfig.secretAccessKey != newKey){
+                awsConfig.secretAccessKey = newKey
+                awsConfigChanged = true
+            }  
+        }
+        if(answers.region){
+            let newRegion = answers.region.trim()
+            if( awsConfig.region != newRegion){
+                awsConfig.region = newRegion
+                awsConfigChanged = true
+            }  
+        }
+        if(awsConfigChanged){
+            let jsonString = JSON.stringify(awsConfig)
+            fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
+            awsConfigFileManager.setNoProfileSync(awsInfo)
+        }
+
+        if(callback){
+            callback(awsDetails)
+        }
+    })
+}
+
+function setProfile(profileName){
+    console.log('Setting awsmobile-cli to use named profile: ' + chalk.blue(profileName))
+    let isSuccessful = awsConfigFileManager.setProfile(profileName) 
+    if(isSuccessful){
+        console.log('Done')
     }else{
-        let awsDetails = resolveAWSConfig()
-        let awsInfo = awsDetails.info
-        let awsConfig = awsDetails.config
+        console.log('Failed to set profile ' + chalk.blue(profileName))
+    }
+}
 
-        inquirer.prompt([
-            {
-                type: 'input',
-                name: 'accessKeyId',
-                message: "accessKeyId: ",
-                default: awsConfig.accessKeyId
-            },
-            {
-                type: 'input',
-                name: 'secretAccessKey',
-                message: "secretAccessKey: ",
-                default: awsConfig.secretAccessKey
-            },
-            {
-                type: 'list',
-                name: 'region',
-                message: "region: ",
-                choices: awsMobileRegions,
-                default: awsConfig.region
-            }
-        ]).then(function (answers) {
-            let awsConfigChanged = false
-            if(answers.accessKeyId){
-                let newKeyId = answers.accessKeyId.trim()
-                if(awsConfig.accessKeyId != newKeyId){
-                    awsConfig.accessKeyId = newKeyId
-                    awsConfigChanged = true
-                }
-            }
-            if(answers.secretAccessKey){
-                let newKey = answers.secretAccessKey.trim()
-                if( awsConfig.secretAccessKey != newKey){
-                    awsConfig.secretAccessKey = newKey
-                    awsConfigChanged = true
-                }  
-            }
-            if(answers.region){
-                let newRegion = answers.region.trim()
-                if( awsConfig.region != newRegion){
-                    awsConfig.region = newRegion
-                    awsConfigChanged = true
-                }  
-            }
-            if(awsConfigChanged){
-                let jsonString = JSON.stringify(awsConfig)
-                fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
-                awsConfigFileManager.setNoProfileSync(awsInfo)
-            }
-
-            if(callback){
-                callback(awsDetails)
-            }
-        })
+function configureAWS(profileName){
+    if(profileName){
+        setProfile(profileName)
+    }else{
+        setConfig()
     }
 }
